Fixed the implicit conversions.

+++ src/utils.h
@@ -12,7 +12,7 @@ void verbose_message(const char *msg);
 void set_verbose_mode(const bool verbose);
 bool verbose_mode();
 char *flat_decode(char *compressed, int size, int &deflated);
-char *deflate(const char *raw, int size, int &writed);
+char *deflate(const char *raw, size_t size, uint32_t &writed);
 string utf16be_to_utf8(string &str);
 string charset_to_utf8(string &str);
